+++ noodles-fastq/src/io/writer.rs
@@ -1,7 +1,11 @@
+//! FASTQ writer.
+
+mod builder;
 mod record;
 
 use std::io::{self, Write};
 
+pub use self::builder::Builder;
 use self::record::write_record;
 use crate::Record;